# openapi/arcade.openapi.yaml
openapi: 3.0.3
info:
  title: Arcade API
  description: |
    REST API for **Arcade** — a Bitcoin SV transaction broadcasting and
    lifecycle-tracking service. Clients submit transactions, look up their
    status, subscribe to status updates (via webhooks or Server-Sent Events),
    and inspect block-processing milestones.

    ## Services and ports

    Arcade is decomposed into independent services, each binding its own HTTP
    port. The endpoints in this document are split across three of them:

    | Service       | Default port | Endpoints                                                                 |
    |---------------|--------------|---------------------------------------------------------------------------|
    | `api-server`  | `8080`       | `/`, `/health`, `/ready`, `/metrics`, `/tx`, `/txs`, `/tx/{txid}`, `/api/v1/*` |
    | `sse`         | `8082`       | `/events`                                                                 |
    | `chaintracks` | `8083`       | `/chaintracks/v1/*`, `/chaintracks/v2/*`                                   |

    Operations that are not served by the default `api-server` declare their own
    `servers` override. Adjust the host/port to match your deployment — in
    production each service sits behind its own Kubernetes Service.

    ## Callback subscriptions

    `POST /tx` and `POST /txs` accept optional `X-CallbackUrl` /
    `X-CallbackToken` / `X-FullStatusUpdates` headers. Supplying a callback URL
...
  version: "1.0.0"
...
servers:
...
security: []
...
paths:
...
  /api/v1/blocks/processing-status:
    get:
      operationId: listBlockProcessingStatus
      tags: [Blocks]
      summary: List block processing milestones
      description: |
        Lists block processing milestones (header seen, `BLOCK_PROCESSED`
        received, compound BUMP built) in descending-height order. Paginates via
        a height-cursor; pass the lowest height returned from the previous page
        as `before-height` to fetch the next page.
      parameters:
        - name: limit
          in: query
          required: false
          description: Page size. Default 50, max 200.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: before-height
          in: query
          required: false
          description: Cursor — pass the lowest height from the previous page.
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        "200":
          description: A page of block processing milestones.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListBlockProcessingStatusResponse"
        "400":
          description: Invalid `limit` or `before-height` query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/blocks/processing-status/{blockHash}:
    get:
      operationId: getBlockProcessingStatus
      tags: [Blocks]
      summary: Get processing status for a single block
      description: Returns the block processing milestones for one block, looked up by block hash.
      parameters:
        - $ref: "#/components/parameters/BlockHashPath"
      responses:
        "200":
          description: Block processing milestones.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockProcessingStatus"
        "404":
          description: Arcade has no record of the block hash.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token shared between the Merkle Service and Arcade. Compared in
        constant time against the configured callback token.
  parameters:
    BlockHashPath:
      name: blockHash
      in: path
      required: true
      description: Block hash (64 hex characters).
      schema:
        type: string
        pattern: "^[0-9a-fA-F]{64}$"
...
  responses:
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    BlockProcessingStatus:
      type: object
      properties:
        blockHash:
          type: string
        blockHeight:
          type: integer
          format: int64
        headerSeenAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
        bumpBuiltAt:
          type: string
          format: date-time
        status:
          type: string
          description: Whether the block is still on the active chain or was orphaned by a reorg.
          enum: [active, orphaned]
        orphanedAt:
          type: string
          format: date-time
        hasBlockProcessed:
          type: boolean
        hasCompoundBUMP:
          type: boolean
      required:
        - blockHash
        - blockHeight
        - status
        - hasBlockProcessed
        - hasCompoundBUMP
    ListBlockProcessingStatusResponse:
      type: object
      properties:
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/BlockProcessingStatus"
        nextCursor:
          type: integer
          format: int64
          description: Lowest height on this page; pass as `before-height` for the next page. Absent on the last page.
      required:
        - blocks
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...